backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pandas
google-generativeai
python-dotenv
sqlalchemy
